Stop treating symbolic binding patterns as constants
This was originally needed to support constant evaluation of name
expressions, but that's now done in a different way.

This is actually a step toward treating all patterns as constants.
The upcoming change will do so in a slightly different way, and so it
will simplify the review to start from a baseline where patterns are
never constant.