Introduce a custom filesystem library
The standard filesystem API lacks significant functionality, ranging
from correct and secure creation of directories and files within them by
using `openat` and avoiding TOCTOU issues, to support for filesystem
locking.

The LLVM filesystem library has more functionality, but uses an API that
is increasingly diverging from the standard, and also fails to defend
against TOCTOU.

This library is designed to carefully model the Unix / Posix filesystem
concepts of `openat` to avoid TOCTOU. However, it also tries to limit
itself to an API subset that LLVM's filesystem library has also
implemneted and so we have a strong reason to expect to be possible to
port to Windows reasonably.

I've also included several benchmarks that show that this implementation
is also dramatically faster for the overwhelming majority of operations
than the C++ standard library. The only places where there is
a regression is in recursively creating directories, and this is
directly connected to the approach used in the C++ standard library not
being secure against TOCTOU.

There are a number of operations not yet included here, I've focused on
a core set of opening, closing, creating, and removing, and then adding
those that I saw the current toolchain code using actively. I'll plan to
expand the operations as needed going forward.