More closely mimic the Clang compilation
Instead of using ASTUnit and tooling APIs, directly mimic what Clang
does during the compilation.

As a result, we now correctly generate definitions for inline C++
functions called by Carbon and functions called recursively by them.

Run the Clang frontend until the translation unit parsing is done, at
which point switch back to Carbon to finish the `Check` face and
interface with Clang through `ASTContext` and `Sema`. In lower, finish
the corresponding Clang compilation phase (CodeGen) and consume the
resulting `llvm::Module`.

Clang does not have the corresponding APIs directly and instead provides
a callback-based mechanism. To map this back to Carbon APIs, we create a
separate thread that gives control back to Carbon through the callbacks,
and then finishes the compilation when the Carbon code is done.

Although essentially a hack, this allows to easily fit into the Carbon
codebase quickly and we have an option of refactoring Clang code in LLVM
upstream in the future to do this directly if this approach proves
fruitful.