Avoid moving around large suspended function states in the deferred definition worklist.
We already go to some effort to avoid moving these, but we end up still
moving them twice: once when adding to the worklist and again when
reversing a chunk of the worklist.

  * To avoid a move when constructing the worklist, add an
    `EmplaceResult` utility that allows the result of a function call to
    be emplaced into a container.
  * To avoid moves when reversing the list, stop reversing it. Instead
    of reversing the list and popping tasks as we run them, we
    accumulate a sequence of tasks for a deferred definition region, run
    them in the order they were enqueued, then pop them all at the end.
    This will in some cases increase the high-water-mark of the size of
    the worklist, but not asymptotically. The same high-water-mark could
    be reached with the old approach by reordering the declarations in
    the source file.

In passing, we no longer create `LeaveDeferredDefinitionRegion` tasks
for non-nested regions. We don't need them, because we can detect that
condition by our reaching the end of the worklist. This means that the
enter / leave region actions are now always in correspondence -- we only
create them for *nested* regions. The tasks have been renamed to convey
this.

We still move the suspended function states around if the worklist grows
to over 64 entries and gets reallocated. We could potentially address
that issue too by switching to a chunked allocation strategy as is used
by `ValueStore` and then make the tasks noncopyable, but I'm not
attempting that in this PR.