Enhance benchmark runner args
There are some very useful default arguments, so teach the runner script
to directly provide them. They can be easily overridden if needed.

Also, tweak the processing of the results to provide a stable order
based on the source order, even when randomized interleaving is enabled.
The randomized interleaving improves the statistical strength of the
benchmarks significantly, but displaying the results in the source order
is much more understandable. This should give roughly the best of both
worlds.