Don't import C++ class definitions until the class is required to be complete.
When importing a class definition, ask Clang to complete it first. Don't
reject importing a class with a virtual base if it's never required to
be complete.

This makes our behavior more consistent with C++'s, and causes class
template specializations to get instantiated as needed when the
type-checking of Carbon requires a C++ class to be complete.

This also removes the recursion from `MapType`, as mapping a class type
no longer maps its definition.