Re-add all that passing around in convert.*
Turns out we did need it - otherwise we end up using the vtable of the
base type (because we're initializing the base subobject). Thought I was
being smart - but past me was smarter.

Wish the test cases were easier to read and I wouldn't've dismissed the
test deltas so easily when the signal gets lost in the noise.