Some fixes to #5914 noticed after merging
Several minor fixes to the content in the `README.md` were pointed out
in review after merging that are fixed here.

I also accidentally undid the removal of the legacy strategy document.
I _think_ I got all of the critical content from this incorporated into
the safety design section, but if there is more that I missed, please
flag and I'm happy to improve that as part of this PR as well.