Rudimentary virtual function call interop support
This only handles the case where the imported C++ type is the first base
class (hasn't derived from other C++ types).

Clang's code for computing virtual function indexes is in Clang's
CodeGenModule (CodeGenVTables, more specifically) which is an
implementation detail of Clang's CodeGen - so I don't think there's
currently a way to reuse that logic from Carbon. So this patch
reimplements that indexing logic in Carbon, which doesn't feel great.

I'm open to suggestions/happy to engage in Clang refactoring if that
seems worthwhile/like a viable path forward, etc.