Improve building of generated sources for ClangD
We have grown more generation rules, so try to use a regex instead of
listing all of them. Also include the runfiles logic that isn't
generated, but needs to be built to be present in the tree.

Also, add a filtering step to remove extra compilations from the
database for the host or "exec" configuration in Bazel. These end up
being duplicate compilations of the same source file. The different
compilation flags would be fine, but they also use a different Bazel
tree for generated files, and one that we can't run a Bazel command to
populate with those generated files easily the way we can and do for
the main configuration. Filtering these out still leaves the main
configuration.

This will likely both make things like `clangd-tidy` even faster by
avoiding redundant work in multiple configurations, and it avoids errors
when those generated files are missing.

The downside of this change is that it means we will need multiple
clangd-tidy actions to build the compilation database in different
configurations in order to see clang-tidy warnings that only occur in
that configuration. This doesn't seem to be a major issue, and we can
pursue it in the future if needed.