Add support for running LLVM optimizer.
Adds a flag `--opt=<mode>` that specifies what to optimize for:

* `--opt=none` turns off the optimizer as much as possible, but still
  respects always_inline.
* `--opt=debug` aims to be the equivalent of `-Og` / `-O1`, and provides
  optimizations that don't affect the ability to debug the program. This
  is the default.
* `--opt=size` optimizes for the size of the produced program, and aims
  to be the equivalent of `-Oz`.
* `--opt=speed` optimizes for the execution time of the produced
  program, and aims to be the equivalent of `-O3`.

Following the approach taken by Clang, the optimization level feeds into
both the configuration of the LLVM pass pipeline and the attributes
added to function definitions generated by the frontend.

Optimization is performed in a new phase, `opt`, which runs between
`lower` and `codegen`.