Use the value representation of T as that of MaybeUnformed(T) where possible.
If the value representation of T is a copy representation, but it copies
all of the bits of T's object representation, then it's OK to use that
as the value representation of MaybeUnformed(T) too.

This fixes the behavior of interop with nullable pointers, which are
represented as an adapter of `MaybeUnformed(T*)`, and need to be passed
to and returned from functions on the Carbon / C++ boundary as `T*`s.