Add a more friendly "latch" synchronization tool
The standard `std::latch` is very restrictive in how it can be used, and
this makes it hard to easily leverage for simple coordination between
a set of dynamically scheduled tasks, where there isn't an interesting
synchronizing "merge" or future result.

This tool makes it easy to establish a latch, hand out handles to it,
and once all are destroyed, take whatever relevant action.