Factor out C-string argv building and simplify vlogs
This extracts the C-string `argv`-like building routine to a more
broadly reusable location. It also sinks the verbose logging logic out
of it and into the relevant runners. In turn, it simplifies the verbose
logging logic significantly.

The biggest functional change is removing the implicit synthesis of
a tool's `-v` verbose flag from the presence of a `vlog` stream.
I thought this would be helpful, but in practice of debugging these
layers it has been more of a hindrance than a help -- I pretty often
only want verbose logging on one side or the other, and we have ways of
explicitly passing a `-v` flag to the underlying tools already. I think
my instinct to do this was just wrong, so rip it out and simplify.

This does add an unused feature -- prepending a prefix of arguments
while building the C-string variant. This isn't used in this PR but will
be used in subsequent PRs and it seemed more disruptive to undo that
logic and then re-do it in a later PR. Let me know if it's too
confusing here.