Cache final impl lookup results.
This allows us to produce the same result for repeated lookups that find
a C++ operator. This isn't a great solution to that problem, as it's not
clear how to extend it to behave correctly across import, but we don't
have a solution for that for C++ interop in general.