Uniform parameter and argument format
This PR consistently formats function parameter and argument lists
as comma-separated lists inside parentheses. In particular, we no
longer place the return parameter outside the parentheses after
a `->` or conditionally omit its name, and we no longer place the
corresponding arguments outside the parentheses after a `to`. This PR
also adds the expression category to the format of parameter insts, so
that they now have the same format as the left-hand side of any other
inst.

The previous parameter format equivocated between a function's return
type and its output parameter pattern, but that equivocation is no
longer viable in the face of `ref` returns and return forms. This PR
resolves the dilemma by treating the `fn` declaration format solely as a
container for the parameter insts, and not as a "signature" for the
function.

For consistency, this PR also removes `to` from the other inst formats
that used it to highlight the output argument.